// frontend/src/api-generated/models/danhMucTaiSanRequest.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */

export interface DanhMucTaiSanRequest {
  /**
     * @minLength 0
     * @maxLength 50
     */
  maDanhMuc?: string;
  /**
     * @minLength 0
     * @maxLength 100
     */
  tenDanhMuc: string;
  moTa?: string;
  /**
     * @minLength 0
     * @maxLength 30
     */
  trangThai: string;
}